Make unify() better match convert(); fix unify bugs.
The top-level unify() function is intended to unify assuming that the type may be converted to the pattern. The original implementation didn't support multiple dereferences of the type. In addition to adding that, this change modifies the code to more closely match the code structure of convert(), and it adds unify() tests for the new typePattern implementation.

As a bonus, the change fixes three long-standing bugs in unify:
1) structs must match field names,
2) unions must match case names,
3) union cases must match typedness.