Rename isRefType to isRefLiteral, add isRefType, use it to fix bug in literal expressions.
isRefLiteral returns whether the type is a reference literal.
isRefType returns whether the type is a reference, following defined types.

There was a bug in literal expressions where isRefLiteral(literalType(T)) (but with isRefLiteral being previously called isRefType) would return false for a reference type. This happened when the type T had no literal type representation. For example if the type was &_, and _ is not a literal, so literalType() returns false.

Instead of using isRefLiteral(literalType(T)), use isRefType(T) which gives the correct result.

This didn't show up in checker unit tests, because the resulting type of the literal was correct by literal.Type(). The issue only showed up in the flowgraph tests when the flowgraph builder would access AddrTypes that weren't actuall addresses.