Convert identifier calls to use type patterns.
This commit demonstrates the initial, primary motivation for type patterns. When checking call arguments, instead of having an all-or-nothing "want type" (being nothing if there is not a common, grounded type for all arguments), we use a type pattern that attempts to generalize the arguments.

The specific win that can be seen in the .pea file changes in this commit is that you no longer need to specify the type of the 1st function argument of a switch call.