Don't add intersection bindings for unreferenced type parameters.
typePatterns tend to be sloppy in that they have type parameters no longer referenced by any variables in their types. Some functions (Switch.sub, for example) assume that the bind set will not have bindings for the unreferenced parameters.

In intersection, before this commit, we created a disjointSet for each type parameter in the list of each type pattern. The result was that every type parameter ended up with a binding in the resulting map even if it was never referenced. This change instead walks the types and adds a disjointSet only for referenced type parameters.