Change convert to try an intersection first instead of last.
The old behavior was to first strip off & and follow defined types and so forth, and only if all of those did not match would it try an intersection. But this means that convert(&int, _) would bind _=int instead of _=&int, which is the incorrect behavior.

This change is to do the intersection first. Only if the intersection fails to we fall back and try to remove & and follow defined types and so forth.