Make it an error for the result of convert() to be a non-grounded type.
It means we could not infer the resulting type of the conversion. It is hard for the caller of convert to check for this condition, because the returned Convert nodes use isect.typ as their type, and whether any of the type variables in isect.typ were bound is then lost. Instead, just make this a requirement of convert, and define it to be an error if the conversion doesn't result in a fully grounded type.