Unify the handling of grounded/ungrounded type patterns when resolving ID calls.
Previously if the pattern was grounded, we would check all candidates to see whether they can convert to it, first explicitly and then if there was still ambiguity, implicitily. But if the pattern was ungrounded, we just filetered out any candidates with non-grounded return types. Now, we use the former logic in both grounded and non-grounded pattern cases.