Change literal implementations to be values, not derefs of a reference.
Before this change, a Lit node in the checker tree always had a reference type and was always wrapped in a deref conversion to make it a value. This used to be needed, because there was no general reference conversion; to make a reference to a literal, the deref would be popped off.

Now there is general reference conversion, so the underlying ref and the deref conversion are unnecessary complexity. Literals can now directly be values.