Share ref and named type following and conversion for literal checking.
This removes a bunch of duplicate code and a bunch of dead code, and reduces uses for functions like isStructRefType, isArrayType, and so forth, which I would like to remove eventually.

This leads to a few semantic changes in the language, allowing more types to be created by literals, like:

	type t &[int]
	…
	x := &t :: [],

which was previous not inferrable, becasue the pattern passed to checkArrayLit was essentially &&[int], and inference of array element type only worked with [T]- and &[T]-style patterns. After this change there can now be any number of references.