Allow types to have the same name if their arity differs.
This is useful to have types with a lower arity that are just like a higher arity but with a fixed argument type. For example:

	Type (K, V) map …
	Type V int_map (int, V) map

	Type (T, E) result …
	Type T result (T, generic_error) result

In addition, this change disallow user defined types to shadow built-in types. For example, it's now an error to try to make a new defined type called `int`.