Add a type name resolution fall back resolving to an imported type with the same name as its module.
If a non-module qualified type name resolves to no types in the current module, check if there is an import with the same name as the type that contains a type with the same name, if so, resolve to that.

This removes module/type name studdering for modules that are intended solely to define a new type. For example, it is no longer necessary to say:

	v := int vector#vector :: new(),

If the vector module is imported under the name `vector`, since it contains a type of the same name, it can be said without the module name:

	v := int vector :: new(),

When checking the current module for a type `vector`, if it fails, it will then look for an import `vector` containing a type `vector`, and resolve to that.