Allow type parameters to be defined in a method's interface constraints.
Previously type parameters were only defined in an function's parameter and return types, no the interface constraints. This was due to a limitation on the previous way that type variables were bound to types, which only allowed binding one way, from a type into a patern. However, the new implementation based on type pattern intersection allows bindings to work both ways, and so we can now support this.

This is useful in many places, in particular in defining functions that operate on generic containers. For example, this is now possible:

	Func sort:(span S) : .length(S)int, [](S, int)&T, [](S, int, int)S, <(T, T)bool

where the type parameter T does not appear in the parameters or return type or sort, but only in the interface.