Require matching ref literal types in order to satisfy interface constraints with an explicit reference literal.
For example, consider:

	Func sort(c C) : [](C, int)&T, .length(C) int, <(T, T)bool

Before this change, <C=string, T=uint8> was a possible binding even though the 2-ary [] function on string returns uint8 and the constraint explicitly states that it is a reference literal &T. But this doesn't make sense for sort — string is immutable, so you can't sort its elements.

The reason it was allowed is that interface constraints could be satisfied by any function as long as the parameter types of the constraint were complicitly convertable to the function's corresponding parameter, and the return type of the function was implicitly convertable to the constraint's return.

This change makes it slightly stricter: the parameter types of the constraint were complicitly convertable to the function's corresponding parameter, however if the constraint's parameter type is explicitly a reference literal, the conversion must not be a dereference—the function must take a reference literal type; and the return type of the function was implicitly convertable to the constraint's return, but if the constraint's return type is explicitly a reference literal, the conversion must not be a reference conversion—the function must return a reference type.

This essentially makes it so that if a function assumes its interface constraints are references, they are only satisfied by functions that indeed are references: the compiler will error out if you try to sort a string.