Fix an error reporting bug when failing to convert a block literal param type.
The code converted the note returned by conventType into an Error. This type conversion works, but the Error type assumes the invariant that its location is never loc.Loc{}, whereas note does not. Some notes had no Loc and it would cause a panic reporting the error. Instead, always ensure we set the location before converting to an Error.