Allow eliding types for function parameters with duplicate types.
If multiple adjacent function parameters have the same type, the type can be elided from all but the last. It is a syntax error for the last parameter not to include a type.

	func f(i, j int) 	// same as func f(i int, j int)
	func g(i, j int, k, l string) 	// same as func g(i int, j int, k string, l string)
	func h(i, j int, k) 	// **syntax error**

For block literals behave similarly, however, there is a special case where it's allowed that no parameters have a type. In this case all types must be inferrable.

	(i, j int){} 	// same as (i int, j int){}
	(i, j int, k, l string){} 	// same as (i int, j int, k string, l string){}
	(i, j int, k){} 	// **syntax error**
	(i, j, k, l){} 	// all parameter types must be inferrable