Add the flowgraph.Unreach instruction; used for !-returning calls.
If a call returns !, LLVM should consider anything after to be unreachable. Without this, LLVM, may fail to see that all paths to a USE are dominated by a DEF, since some paths would, without the unreachable annotation, appear to be valid. This happens, for example, when one branch of a case-call is a !-returning call, but the other branch returns a value. After optimizing the flowgraph may create a graph like this:

	branch 1	2
1:
	DEF A
	JMP 3
2:
	CALL !-RETURNING FUN
	JMP 3	// But really this is unreachable
3:
	USE A

Where LLVM, if not told that the JMP 3 in basic block 2 is really unreachable, may detect as an invalid USE A, that is reachable through 2, where A is not defined.