Remove a singleInit loop that can never loop in isBlockLiteralArg.
This loop can never do anything but immediately break, because singleInit(v) always returns nil on its very first call. The reason is that the v argument is guaranteed to be used by a call (it is only called with a Value that is the argument of a call), and singleInit always returns nil if its argument is used by a Call.