Speed up llvm tests using llc and clang instead of llvm-as, opt, llc, and clang.
The previous steps to compile the test .ll files was:
	use llvm-as to generate a .bc file,
	use opt to optimize the .bc file,
	use llc to compile the .bc file to a .s file,
	use clang to compile and link the .s file
Instead, we now just use:
	use llc to compile the .ll file to an .o file,
	clang to link the .o file.

Before this change here is the time to run llvm tests on my laptop:

	% time go test ./llvm -count=1
	ok  	github.com/eaburns/pea/llvm	7.667s
	        7.83 real         6.49 user         2.76 sys
	% time go test ./llvm -count=1
	ok  	github.com/eaburns/pea/llvm	7.679s
	        7.84 real         6.50 user         2.79 sys
	% time go test ./llvm -count=1
	ok  	github.com/eaburns/pea/llvm	7.520s
	        7.67 real         6.50 user         2.77 sys

Here is the time after this change:

	% time go test ./llvm -count=1
	ok  	github.com/eaburns/pea/llvm	3.163s
	        3.31 real         4.37 user         1.72 sys
	% time go test ./llvm -count=1
	ok  	github.com/eaburns/pea/llvm	3.465s
	        3.61 real         4.38 user         1.71 sys
	% time go test ./llvm -count=1
	ok  	github.com/eaburns/pea/llvm	3.201s
	        3.35 real         4.39 user         1.75 sys

Note that we could use just clang to compile and link the .ll file. But it seems that the clang install used for github.com actions doesn't compile .ll files. Also, the current two-step approach is in line with how peac compiles .ll files, and it seems to be just as fast anyway.