Optimize switch on the result of a built-in <=> into comparison operators.
A complete switch function (less?equal?greater?) with a built-in <=> operator where all the block arguments only return constant booleans can be optimized into a single comparison call.

For example:

	1 <=> 2 less? { true } _? { false }

is the same as

	1 < 2

This optimization is useful so that we can remove the built-in comparison operators infavor of a library implementation based on <=> while not losing any performance.