Simplify call substitution for interface constraints.
This used to wrap the called constraint function in a block, because a long time ago this was the only way to add a possible reference conversion to the return type of the call. But Ref conversions were added as a conversion type since then, so this wrapping is no longer needed.

An additional benefit of this simplification is that the flowgraph builder can better find optimizations in functions with interface constraints, since it doesn't need to look into a wrapping block to see what built-in function is being called.