Use clang -O2 instead of llc to compile .ll files.
This does two things: 1) it applies clang's -O2 optimization pipeline, and 2) as a requisite for that, it sets the target datalayout and triple values for the target arch to match clang's, which is needed to properly optimize and also probably needed to communicate with some clang-compiled C.

The optimizations give pretty good performance on the regexp unit tests.

On my macbook air the results are:

With llc:
	%  time ./modules/regexp/regexp.test
	…
	        9.17 real         7.74 user         1.00 sys
	%  time ./modules/regexp/regexp.test
	…
	        9.06 real         7.71 user         0.93 sys
	% time ./modules/regexp/regexp.test
	…
	        9.17 real         7.74 user         0.98 sys

With clang -O2:
	% time ./modules/regexp/regexp.test
	…
	        5.42 real         4.04 user         0.81 sys
	% time ./modules/regexp/regexp.test
	…
	        5.38 real         4.04 user         0.75 sys
	…
	        5.47 real         4.07 user         0.84 sys

With clang -O3:
	% time ./modules/regexp/regexp.test
	…
	        5.49 real         4.00 user         0.89 sys
	% time ./modules/regexp/regexp.test
	…
	        5.35 real         4.00 user         0.79 sys
	% time ./modules/regexp/regexp.test
	…
	        5.42 real         4.04 user         0.80 sys

Decided to go with -O2, since it's about as performant as -O3 anyway, and less processed.