Replace named bool and ordering built-in types with union literals.
First, these two types were somewhat exceptional becasue they were the only built-in types that behaved as defined types (defined with an underlying union literal).

Second, ordering is insufficient, because floats aren't ordered, but partially ordered; they need a new type for their <=> operator. Instead of introducing yet another special case for a partial_ordering type, let's just remove the special cases and use union literals directly.

Since no one wants to actually type out [false?, true?] everywhere, the library can define bool as an alias. Note that it is an alias, not a new type, so that it behaves well with the built-in functions returning the union literal. Specifically:

	Func &&(a bool, b (){bool}) bool

shoudl be able to work on:

	x && {1 = 2}

where the second argument type is (){[false?, true?]), since = is the built-in = that returns [false?, true?]. If bool were a type instead of an alias: (){[false?, true?]} would not convert to (){bool}, and we would need a second definition of && to work on the defined bool type.