Remove the need for llvm-ar in libpea/build.sh.
We were using llvm-ar since it had the ability to append the contents of one .a file to another. But in an effort to remove the need for special llvm dependencies, this change just re-implements that by extracting the archives into a temp directory, and building a new archive with all of their contents.