Change built-in [] and .length functions to be defined on only literal arrays and string.
Previously they were defined for array literals, strings, or any non-opaque defined type that could be determined to have an underlying array or string definition.

Nevertheless, this change makes those built-in functions defined only for literal array types and the type string. This simplifies the implementation of them in the checker quite a bit. It also simplifies how they are specified in English prose too.

There seems to be only two ramifications of this, and both seem acceptible:

1) You cannot implicitly convert an explicit conversion, so this pattern no longer works:

	(my_array :: [1, 2, 3])[2]

But it doesn't seem like a useful pattern outside of unit tests for the checker anyway.

2) Defined string types no longer have built-in [] and .length functions. This seems reasonable, since those functions are trivial to define anyway.