Change built-in selector functions to be defined only on struct literal reference types.
Before this change any type that was a struct or reference to a struct had built-in selector functions for all of its fields. Now, instead, the built-in functions are only defined on literal types.

Since all the previous types can implicitly convert to the equivalent struct literal, in all practical cases this change doesn't change much. But it makes built-in selectors inline with built-in [], which are defined on literal arrays, and it makes the description of what types have built-in selector functions defined much simpler.

There is one thing made now illegal by this change, and that is using a selector on an explicit conversion to a named type is no longer possible, since you cannot implicitly convert an explicit conversion:

	type point [.x int, .y int]
	// Can't implicitly convert explicit conversion
	// (point :: [.x 5, .y 6]) to &[.x int, .y int].
	// This is now illegal.
	(point :: [.x 5, .y 6]).y

But since that doesn't seem like a useful thing to do anyway, it's fine.