Change built-in switch functions to be defined only on union literal reference types.
Before this change any type that was a union or reference to a union had built-in switch functions for all combinations of its cases. Now, instead, the built-in functions are only defined on literal types.

This is similar to the previous change which changed selector functions, but for switch functions instead.