Change built-in bit-wise and arithmetic operators to be defined only on built-in types.
Before this change the bit-wise operators were defined for all types with an underlying integer type, even user-defined named types. Similarly for the arithmetic operators and underlying numeric types. This change makes it so these functions are only defined on the built-in integer and numeric types instead.

This simpifies the implementation: we can enumerate the built-in functions. And it is still possible for a user-defined type to implement these in a library, if a bit annoying.