Change convert() to return a type pattern.
Previously convert() returned an error if the resulting conversion was not a grounded type. The resulting type was the .Type() of the returned *Convert node.

This change allows convert() to succeed even when the result is not grounded. The caller can look at the resulting pattern to decide whether it is an error or not.

This is in preparation for an upcoming change where resolveIDCall will do the conversion check for function return values _first_ instead of _last_. This will allow the desired pattern of the call's return value to be substituted when checking argument expressions. But to do this, we will need to allow the conversion to succeed even if the result is not a grouded type. For example, the desired pattern may be _ and the return value of the function may also be _ or some other non-grounded pattern. We should allow this, with the assumption that parameter and iface constraint substitution will bind all the type varaibles.