Change Func.sub to take as slice of TypeParms bound to TypeVars in the map.
Currently all values in the substitution map are fully-grounded types. But in a subsequent change this will become untrue. In that case, we need a way to pass to Func.sub() a list of TypeParms that are bound to TypeVars in the values of the substitution map so that these TypeParms can be added to the Func's type parameter set.