Add if:-style switch functions for union literal types.
It is the same as the ?-style, but the return type, and thus the result type of all function arguments is [.]. This is convinient in cases where you don't actually have a return value. The ?-style doesn't work well here, because it requires you to `[.] ::`-convert the last expression of the first argument in order to explicitly infer the `[.]` type for all remaining function arguments.