Better ID call resolution by passing whether the return is being explicitly converted.
Before this change, there was a two step process to check candidate return values. First we filtered out any that could not even explicitly convert — these can never be used, since even an explicit conversion cannot make the types work. Second, if there was still ambiguity, we filtered out any that could not implicitly convert. This behavior is bad for two reasons.

First, if the parent expression is an explicit conversion, it could be that we have multiple candidates that would work, but we would throw out any that only work due to explicit conversion. But the conversion _is_ explicit; I would expect this to be ambiguous. I would not expect it to select one or the other.

Second, and worse, if the parent is an explicit conversion and two candidates _can_ explicitly convert, but neither can implicitly convert, we should not discard _either_ of them simply due to lack of implicit conversion. The conversion is _not_ implicit; it is being expressed explicitly. We should accept both candidates. It will either be ambiguous—fine, it would have failed anyway on the old rules, but also it may be that only one candidate has satisfied interface constraints, in which case we should use that one.

This change simply passes a boolean to resolveIDCall which says whether the parent is an explicit conversion or not. We use that to check whether the return value can be converted, and don't do the two step check at all.