SubExpr parsing prefers FuncName over Expr.
The only change this makes is that ([]) is now interpreted as a [] identifier, referring to a function named [], instead of an empty array literal.

This is preferred, because there is not much need to put parentheses around an array literal, but without this change there is no convenient way to reference an index or slice function by name.

Note, that before you could do: (x, y){ x[y] } and (x, y, z){ x[y, z] } but there was no way to ever call a unary or N>3-ary [] function. Now, such a function may be of dubious use, but we do allow defining it, so we should allow calling it.