Simplify argument and return value error messages.
Instead of adding an extra level of nesting stating that the argument or return failed to convert, just return the underlying conversion error. For arguments it will have the location of the expression that failed the conversion, which is exactly where you want to go, and for the return value it will have the location of the call that failed.