Refactor the checker type into topScope, above all scopes, instead.
Previously, checker held state for Check, but it was only available to the Check function itself, and not any of the recursive calls. In a future change, this state will be used in building better error messages, since much of the state is about error locations, verbosity, depth, and path trimming. Instead of plumbing *checker throughout every function, this change uses the fact that a scope is already plumbed through. By replacing the checker with topScope, it's always available to functions with a scope by simply going to the top scope to get it.