Change tracking of required reference literal parameters in FuncDecls.
A parameter only accepts a reference literal if the defining constraint type was a reference literal. Through substitution a FuncDecl may end up with a parameter that was originally not a reference literal, but was substituted with a reference literal. We need a separate way to track which parameters were _defined_ as a reference literal (as opposed to just substituted with one).

Before this change, it was tracked by passing around both the *FuncDecl and the original FuncDef's version of the same *FuncDecl (or optionally nil). This was a bit clumsy, since it 1) it was confusing to have two *FuncDecls passed around, and 2) it required passing the index of a constraint around so its original defining counterpart could be found.

This change instead tracks an explicit array for each parameter and the return type of a FuncDecl indicating whether it must be a reference literal.