Properly enforce reference literal constraints when the literal is an iface type arg.
If a constraint has a reference literal, it is enforced that the satisfying function must have a reference literal too. Before this change, there was a bug that a constraint coming from an interface, instantiated with a reference literal type argument did not apply the restriction.

This is important for the following exmaple:

	Iface (S, T) span {
		.length(S)int,
		[](S, int)T,
		[](S, int, int)S,
	}

For this interface, we want to distinguish (S, T) span from (S, &T) span, where the latter requires the [](S, int)&T function to return a reference literal. This way, the constrained function can assume the result is assignable, whereas for (S, T) span, it would not assume it could usefully assign to the return of [](S, int)T.