Add user-defined :: functions for explicit conversions.
Conversions, ::, are now allowed as interface constraints.

For :: expressions and :: constraints, first the search looks for user-defined :: functions where the arg and return can implicitly convert to the appropriate types. If theer is exactly 1, it is used as the conversion (possibly shadowing a built-in conversion). If there are more than one, it is an ambiguous error. If there are 0 user-defined conversions, then it tries a built-in conversion; possibly leading to a not-found error if the built-in conversion is not possible.