Prefer implicit convert over user-defined :: functions.
If a type can be implicitly converted, prefer that over user-defined :: functions. This allows to use :: as a constraint that accepts the conversion target type as the source type without calling the user-defined :: to "convert" the type.

This is more natural than calling the user defined :: in useful cases involving "virtual types".

Consider:

	// A virtual collection has a .for_each function
	// that calls the for:each: function of a captured collection type.
	Type T collection [.for_each ((T){}){}]

	// Calls f for each T in c.
	Func for:each:(c T collection, f (T){}) { c.for_each(f) }

	// Converts anything with a for:each: on Ts into a T collection.
	Func ::(c C : for:each:(C, (T){})) T collection

	// Returns f(c).
	Func ->(c C : ::(C)T collection, f (T collection){U}) U

It is useful to be able to pass anything convertable to a T collection as the left-hand side of ->. A T collection itself can satisfy the constraints of the user-defined :: function. But this would needlessly re-wrap it in another T collection. Instead, after this change, it would be implicitly converted instead of user-defined-:: converted.