Make << and >> take the same operand for both arguments.
Previously, the second argument was always int. This conflicts with LLVM, which requires them to be the same type. We could change the LLVM code generator to deal with int, but it's a hassle and possibly less efficient in some cases. But further, there is no real disadvantage for using the same type for both arguments here. Using int for the second argument in practice just made a weird special case.