Remove built-in incomplete ?-style switch functions.
Now that we have if:-style switches, there is not real need for built-in ?-switches that are incomplete. Previously these would behave just like their incomplete, if:-style counterpart. But the problem is that it prevented some useful cases for incomplete ?-style funcitons. For example,

	// x := int_option none? { 0 },
	// y := string_option none? { "" },
	// z := point_option none? { panic("impossible") },
	Func none?(o T option, f (){T}) T {
		return: o some? (t){ t } none? f
	}

Which is is a convenient way to get the some?-value of an option or a default in the none? case.

Before this change to remove the built-in none? function, any such call would instead be ambiguous.