Add a comment pointing out misleading var def syntax.
Var def syntax is:
	var x := some_type :: some_expr

But note that today the `some_type ::` part is not a conversion.
It is part of the var def syntax; it just _looks_ like a conversion.
That means that some_expr is _implicitly_, not explicitly converted to some_type.
This is confusing. It should just be explicit.