Remove a broken unit test.
It was trying to test that new(int, T)[T] could not be assigned to a named type.
But it can!
The reason the test passed is that the types were mixed up,
and it was trying to assign new(5, 0) of type [int] to
the variable byte_array_ref of type [&int],
and [int] cannot convert to [&int].