Don't bother filtering type parms in makeTypePattern.
Before this change, it would filter the type parameters to only those bound to a type variable in the type. Instead, don't bother. The number of type parameters is not going to be so large that the filtering matters. And when printing multiple type patterns as strings, keeping the type parameters the same makes the names more consistent.