Add implicit 0-ary function conversion.
A 0-ary function can be passed as a function with any number of parameters—the parameters are just ignored. This allows some convenient patterns, like passing a 0-ary function to for:to:do: in cases where you don't care about the index:

	for: 1 to: 10 do: {
		// Something that doesn't care about the index.
	}

More importantly, this will enable a change to unions: today cases with no type are untyped, but it would simplify several things if instead all cases were typed. Untyped cases will instead be [.]-typed. This is similar to how functions with no return aren't untyped, but are [.] typed.

However, it would be really clumsy if we required passing ([.]){T} typed function arguments:

	x true? (_ [.]){
		// …
	} false? (_ [.]) {
		// …
	}

This conversion allows to continue passing 0-ary functions here.