Add first-class support for type pattern testing in parser and checker.
Testing with type patterns used to be quite hacky, involving rewriting the type pattern source into a type and pulling parameters out into the pattern. It meant that sharing type parameters and type variables among patterns in tests was difficult, and string representation of these patterns leaked implementation details (the type parms were secretely named Z in the implementation).

Instead, this change adds ParseTypePattern which is just like ParseType, but allows type variables to be named ? — which is not official Pea syntax, but test-only syntax. And the checker now adds type parameters for tests to the module scope, so they can be shared among test type patterns easily.